requirements.txt
python-memcached
celery==4.1.0
redis==2.10.6
